Change DownloadCount type to ulong

+++ Reportably/Reportably.Data/Migrations/20200210120917_Initial.cs
@@ -60,7 +60,7 @@ namespace Reportably.Data.Migrations
                     IsDeleted = table.Column<bool>(nullable: false),
                     DeletedOn = table.Column<DateTime>(nullable: true),
                     UploadedFile = table.Column<Guid>(nullable: false),
-                    DownloadCount = table.Column<long>(nullable: false)
+                    DownloadCount = table.Column<decimal>(nullable: false)
                 },
                 constraints: table =>
                 {
